Refactor User model schema
This commit refactors the User model schema in the server/models/User.js file. It updates the import statements to use double quotes instead of single quotes. It also adds missing commas after the trim property in the username, firstName, and lastName fields. Additionally, it adds a minlength property to the password field. The groups, games, castMembers, and joinedGames fields are updated to use an array of objects with a type and ref property. Finally, the pre-save middleware function is updated to use double quotes instead of single quotes.